Renames paths to remove underscores.
