+++ uninstall.php
@@ -10,7 +10,7 @@ if ( ! defined( 'WP_ROCKET_CONFIG_PATH' ) ) {
 	define( 'WP_ROCKET_CONFIG_PATH', WP_CONTENT_DIR . '/wp-rocket-config/' );
 }
 
-require_once dirname( __FILE__ ) . '/inc/classes/WP_Rocket_Uninstall.php';
+require_once dirname( __FILE__ ) . '/inc/Engine/WPRocketUninstall.php';
 
-$rocket_uninstall = new WP_Rocket_Uninstall( WP_ROCKET_CACHE_ROOT_PATH, WP_ROCKET_CONFIG_PATH );
+$rocket_uninstall = new WPRocketUninstall( WP_ROCKET_CACHE_ROOT_PATH, WP_ROCKET_CONFIG_PATH );
 $rocket_uninstall->uninstall();
